[resto druid] First steps of directory refactor
