+++ src/Parser/Druid/Restoration/Modules/Features/AverageHots.js
@@ -5,7 +5,7 @@ import SpellIcon from 'common/SpellIcon';
 import SPELLS from 'common/SPELLS';
 import Analyzer from 'Parser/Core/Analyzer';
 import Combatants from 'Parser/Core/Modules/Combatants';
-import Mastery from './Mastery';
+import Mastery from '../Core/Mastery';
 
 class AverageHots extends Analyzer {
   static dependencies = {
