+++ src/Parser/Druid/Restoration/Modules/Talents/CenarionWard.js
@@ -6,7 +6,7 @@ import SpellIcon from 'common/SpellIcon';
 import SPELLS from 'common/SPELLS';
 import Analyzer from 'Parser/Core/Analyzer';
 import Combatants from 'Parser/Core/Modules/Combatants';
-import Mastery from './Mastery';
+import Mastery from '../Core/Mastery';
 
 class CenarionWard extends Analyzer {
   static dependencies = {
