+++ src/Parser/Druid/Restoration/Modules/Talents/SpringBlossoms.js
@@ -7,7 +7,7 @@ import SpellLink from 'common/SpellLink';
 import SPELLS from 'common/SPELLS';
 import Analyzer from 'Parser/Core/Analyzer';
 import Combatants from 'Parser/Core/Modules/Combatants';
-import Mastery from './Mastery';
+import Mastery from '../Core/Mastery';
 
 class SpringBlossoms extends Analyzer {
   static dependencies = {
