+++ src/Parser/Druid/Restoration/README.md
@@ -9,16 +9,16 @@
 | [Efflorescence uptime](https://github.com/buimichael/RestoDruidAnalyzer/blob/master/src/Main/Parser/Modules/Features/Efflorescence.js) | Get uptime on efflorescence | 90-100% |
 | [Regrowth/clearcasting usage](https://github.com/buimichael/RestoDruidAnalyzer/blob/master/src/Main/Parser/Modules/Features/Clearcasting.js) | Tracks clearcasting and regrowth usage. | 90-100% |
 | [Cast Efficiency](https://github.com/buimichael/RestoDruidAnalyzer/blob/master/src/Main/CastEfficiency.js) | Displays the amount of casts and estimates the max amount of casts possible to determine your skill at keeping important spells (close to) on cooldown. The estimated max amount of casts is slightly inaccurate due to not considering Haste increasers like Bloodlust and boss abilities. | >90% |
-| [Drape of Shame](https://github.com/buimichael/RestoDruidAnalyzer/blob/master/src/Main/Parser/Modules/Legendaries/DrapeOfShame.js) |  | 100%|
-| [Prydaz, Xavaric's Magnum Opus](https://github.com/buimichael/RestoDruidAnalyzer/blob/master/src/Main/Parser/Modules/Legendaries/Prydaz.js) | | 100% |
-| [Velen's Future Sight](https://github.com/buimichael/RestoDruidAnalyzer/blob/master/src/Main/Parser/Modules/Legendaries/Velens.js) | | 100% |
-| [Dark Titan's Advice](https://github.com/buimichael/RestoDruidAnalyzer/blob/master/src/Main/Parser/Modules/Legendaries/Velens.js) | 100% of random bloom effect and 300% of the non overhealing part of manual bloom| 100% |
-| [Ekowraith](https://github.com/buimichael/RestoDruidAnalyzer/blob/master/src/Main/Parser/Modules/Legendaries/Velens.js) | Increased effect from ysera's gift and if specced into guardian affinity the damage reduction is translated into throughput aswell | 80-90% |
-| [Essence of Infusion](https://github.com/buimichael/RestoDruidAnalyzer/blob/master/src/Main/Parser/Modules/Legendaries/Velens.js) | | 100% |
-| [Sephuz's Secret](https://github.com/buimichael/RestoDruidAnalyzer/blob/master/src/Main/Parser/Modules/Legendaries/Velens.js) | The average haste gain from both status and procc and calculate throughput. This assumes that 1 haste rating = 1 int.| 90% |
-| [Tearstone of Elune](https://github.com/buimichael/RestoDruidAnalyzer/blob/master/src/Main/Parser/Modules/Legendaries/Velens.js) | Counts the amounts of free rejuvs gained (and procchance). Calculates throughput by checking the throughput from 1 rejuvenation and multiplying with the amounts of free rejuvenations gained. | 70-90% |
-| [X'oni's Caress](https://github.com/buimichael/RestoDruidAnalyzer/blob/master/src/Main/Parser/Modules/Legendaries/Velens.js) | The healing contributed by the iron bark effect. Doesn't take into consideration of the reduced iron bark CD. | 90% |
-| [Soul of the Archdruid / Soul of the Forest](https://github.com/buimichael/RestoDruidAnalyzer/blob/master/src/Main/Parser/Modules/Legendaries/Velens.js) | Tracks the abilities consumed by the SotF buff and the healing done from the healing increases. | 70-100% |
+| [Drape of Shame](https://github.com/buimichael/RestoDruidAnalyzer/blob/master/src/Main/Parser/Modules/Items/DrapeOfShame.js) |  | 100%|
+| [Prydaz, Xavaric's Magnum Opus](https://github.com/buimichael/RestoDruidAnalyzer/blob/master/src/Main/Parser/Modules/Items/Prydaz.js) | | 100% |
+| [Velen's Future Sight](https://github.com/buimichael/RestoDruidAnalyzer/blob/master/src/Main/Parser/Modules/Items/Velens.js) | | 100% |
+| [Dark Titan's Advice](https://github.com/buimichael/RestoDruidAnalyzer/blob/master/src/Main/Parser/Modules/Items/Velens.js) | 100% of random bloom effect and 300% of the non overhealing part of manual bloom| 100% |
+| [Ekowraith](https://github.com/buimichael/RestoDruidAnalyzer/blob/master/src/Main/Parser/Modules/Items/Velens.js) | Increased effect from ysera's gift and if specced into guardian affinity the damage reduction is translated into throughput aswell | 80-90% |
+| [Essence of Infusion](https://github.com/buimichael/RestoDruidAnalyzer/blob/master/src/Main/Parser/Modules/Items/Velens.js) | | 100% |
+| [Sephuz's Secret](https://github.com/buimichael/RestoDruidAnalyzer/blob/master/src/Main/Parser/Modules/Items/Velens.js) | The average haste gain from both status and procc and calculate throughput. This assumes that 1 haste rating = 1 int.| 90% |
+| [Tearstone of Elune](https://github.com/buimichael/RestoDruidAnalyzer/blob/master/src/Main/Parser/Modules/Items/Velens.js) | Counts the amounts of free rejuvs gained (and procchance). Calculates throughput by checking the throughput from 1 rejuvenation and multiplying with the amounts of free rejuvenations gained. | 70-90% |
+| [X'oni's Caress](https://github.com/buimichael/RestoDruidAnalyzer/blob/master/src/Main/Parser/Modules/Items/Velens.js) | The healing contributed by the iron bark effect. Doesn't take into consideration of the reduced iron bark CD. | 90% |
+| [Soul of the Archdruid / Soul of the Forest](https://github.com/buimichael/RestoDruidAnalyzer/blob/master/src/Main/Parser/Modules/Items/Velens.js) | Tracks the abilities consumed by the SotF buff and the healing done from the healing increases. | 70-100% |
 | [Dreamwalker](https://github.com/WoWAnalyzer/WoWAnalyzer/tree/master/src/Parser/RestoDruid/Modules/Features/Dreamwalker.js) | Calculates the amount healed by Dreamwalker. | 100% |
 The accuracy percentages assume there are no bugs in the implementation, the accuracy as of right now is not yet verified.
 ## Suggestions
@@ -45,4 +45,4 @@ This may be outdated.
 | Avergae mana spent during innervate | <220k | <180k | <130k |
 | Seconds capped on mana during innervate | N/A | N/A | >0 |
 
-* Spells: Tranquility, Innervate, Essence of G'hanir, Cenarion Ward, Flourish, Tree Of Life, Arcane torrent, Velens, 
+* Spells: Tranquility, Innervate, Essence of G'hanir, Cenarion Ward, Flourish, Tree Of Life, Arcane torrent, Velens,
